+++ backend/lighthouse/src/main/java/com/ssafy/lighthouse/domain/study/entity/StudyLike.java
@@ -12,7 +12,7 @@ import javax.persistence.*;
 @RequiredArgsConstructor
 public class StudyLike extends BaseEntity {
     @NonNull
-    private int studyId;
+    private Long studyId;
     @NonNull
-    private int userId;
+    private Long userId;
 }
