+++ backend/lighthouse/src/main/java/com/ssafy/lighthouse/domain/study/repository/SessionRepository.java
@@ -10,5 +10,5 @@ import com.ssafy.lighthouse.domain.study.entity.Session;
 
 public interface SessionRepository extends JpaRepository<Session, Integer> {
 	@Query(value = "SELECT s FROM Session s WHERE  s.isValid = 1 AND s.studyId = :studyId ORDER BY s.seqNum")
-	List<Session> findByStudyId(@Param("studyId") int studyId);
+	List<Session> findByStudyId(@Param("studyId") Long studyId);
 }
